Getting the positions
Using pointers for getting the positions of the boats. Two functios have been created, the first one is called "positions" which is responsible for reading the user inputs, the second  function named "repeat" is used for detecting whether the user typed any repeated value. The last one still need to be completed since the choosen positions are not set on the board yet.

The user is already capable of choosing  the spaces, nevertheless the choosen spaces are not used for setting any value on the board so it's not possible to make comparisons and find already choosen spaces and void spaces on the board.