Finished the destroyer func and solved some bugs
Solved a couple bugs on the "destroyer" function. After solving them, changed some parts of the "tugShip" function so I can have more or less the same code in both functions, therefore this code can be further substituted in another separate function, since they are equals, but used in different parts of the program.

Some minor changes in the code have been made, such as renaming some variables and creating a header file which will be used later.