variables header, computer_game files, and lockBoats function
Created computer_game.h and .cpp, these files will be used for generating the positions for the computer automatically.

As some global variables are used, it is better to have a header if some, of them. lockBoats help to aislate each boat so the program can diferentiate boats and does not see parts declared in different boats as parts of the same boat.