Changes in the 'Boats' Class
A few changes on how the program create and use the objets, before changing, boat objects where created and destroyed after the boats where placed on the board, now the boats are created but not destroyed after placing them on the boart, the program works in part, there are a few bugs to be corrected though.

Also, another class responsible for creating linked lists were created, but it is not being used so far. First I need to correct the bugs in the later changins I made in the code, then I'll implement that class