Getting address of the boats
Now in the Boats class there are variables and methods to get the address of all the different boats in the board separately. The addresses now are a member of the classes and may allow better ways of checking and seeking for a boat in the board, as well as a more efficient way of settling the boats up in the board and avoiding too many loops.